refactor: inline tx translation helpers
