+++ packages/i18n/src/translation-functions/internal/createLookupOptions.ts
@@ -1,0 +1,17 @@
+import { NormalizedLookupOptions, LookupOptionsFor } from '../types/options';
+import type { DataFormat } from '@generaltranslation/format/types';
+
+/**
+ * Add the default format to caller-provided lookup options.
+ */
+export function createLookupOptions<T extends DataFormat>(
+  locale: string,
+  options: LookupOptionsFor<T>,
+  defaultFormat: T
+): NormalizedLookupOptions<T> {
+  return {
+    ...options,
+    $format: (options.$format ?? defaultFormat) as T,
+    $locale: locale,
+  };
+}
