+++ packages/i18n/src/translation-functions/t.ts
@@ -1,9 +1,14 @@
-import {
-  resolveStringContent,
-  resolveStringContentWithFallback,
-} from './internal/helpers';
+import { getI18nCache } from '../i18n-cache/singleton-operations';
+import { getI18nConfig } from '../i18n-config/singleton-operations';
 import { GTTranslationOptions } from './types/options';
 import { getLocale } from '../helpers/locale';
+import { interpolateMessage } from './utils/interpolation/interpolateMessage';
+import { createLookupOptions } from './internal/createLookupOptions';
+import type {
+  StringContent,
+  StringFormat,
+} from '@generaltranslation/format/types';
+import type { LookupOptionsFor } from './types/options';
 
 /**
  * Translate a message
@@ -91,3 +96,76 @@ export type TemplateSyncResolutionFunction = (
   strings: TemplateStringsArray,
   ...values: unknown[]
 ) => string;
+
+/**
+ * Just do a simple lookup of the translation
+ * And interpolate
+ */
+export function resolveStringContent(
+  locale: string,
+  content: StringContent,
+  options: LookupOptionsFor<StringFormat> = {}
+): StringContent | undefined {
+  const i18nCache = getI18nCache();
+  const lookupOptions = createLookupOptions(locale, options, 'STRING');
+  const translation = i18nCache.lookupTranslation(
+    lookupOptions.$locale,
+    content,
+    lookupOptions
+  );
+  if (translation == null) return undefined;
+  return interpolateMessage({
+    source: content,
+    target: translation,
+    options: lookupOptions,
+    sourceLocale: getI18nConfig().getDefaultLocale(),
+  });
+}
+
+/**
+ * Lookup translation, fallback to source
+ */
+export function resolveStringContentWithFallback(
+  locale: string,
+  content: StringContent,
+  options: LookupOptionsFor<StringFormat> = {}
+): StringContent {
+  const i18nCache = getI18nCache();
+  const lookupOptions = createLookupOptions(locale, options, 'STRING');
+  const translation = i18nCache.lookupTranslation(
+    lookupOptions.$locale,
+    content,
+    lookupOptions
+  );
+  return interpolateMessage({
+    source: content,
+    target: translation,
+    options: lookupOptions,
+    sourceLocale: getI18nConfig().getDefaultLocale(),
+  });
+}
+
+/**
+ * Lookup translation
+ * fallback to runtime translate
+ * Fallback to source
+ */
+export async function resolveStringContentWithRuntimeFallback(
+  locale: string,
+  content: StringContent,
+  options: LookupOptionsFor<StringFormat> = {}
+): Promise<StringContent> {
+  const i18nCache = getI18nCache();
+  const lookupOptions = createLookupOptions(locale, options, 'STRING');
+  const translation = await i18nCache.lookupTranslationWithFallback(
+    lookupOptions.$locale,
+    content,
+    lookupOptions
+  );
+  return interpolateMessage({
+    source: content,
+    target: translation,
+    options: lookupOptions,
+    sourceLocale: getI18nConfig().getDefaultLocale(),
+  });
+}
